requirements.txt
openai
slack-bolt
slack-sdk
langchain
sentence-transformers
GitPython
pymilvus
llama-index-core
llama-index-readers-web
llama-index-readers-github
llama-index-vector-stores-milvus
llama-index-embeddings-langchain
llama-index-postprocessor-cohere-rerank
llama-index-llms-langchain
llama-index-llms-fireworks
ipython
cohere
trafilatura
nbconvert
langchain-openai
upstash-redis
jupyter-notebook-parser
requests==2.31.0
youtube-transcript-api
